+++ store/index.ts
@@ -1,9 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 import imageReducer from "./slices/imageSlice";
+import sliceReducer from "./slices/sliceSlice";
 
 const store = configureStore({
   reducer: {
     image: imageReducer,
+    slice: sliceReducer,
   },
 });